storageccl: Implement lease preferences as a CCL feature
Release note (enterprise change): The ability to set lease placement
preferences in zone configs has graduated from experimental to fully
supported. Existing lease placement preferences will continue to
function as in v2.0.
