+++ pkg/cli/cli_test.go
@@ -662,13 +662,7 @@ func Example_zone() {
 	// zone set system.jobs@primary --file=./testdata/zone_attrs.yaml
 	// pq: setting zone configs on indexes or partitions requires a CCL binary
 	// zone set system --file=./testdata/zone_attrs_advanced.yaml
-	// range_min_bytes: 1048576
-	// range_max_bytes: 134217728
-	// gc:
-	//   ttlseconds: 90000
-	// num_replicas: 3
-	// constraints: {+region=us-east-1: 1, '+zone=us-east-1a,+ssd': 1}
-	// experimental_lease_preferences: [[+region=us-east-1], [+zone=us-east-1a]]
+	// pq: setting lease placement preferences in zone configs requires a CCL binary
 	// sql -e create database t; create table t.f (x int, y int)
 	// CREATE TABLE
 	// zone set t --file=./testdata/zone_range_max_bytes.yaml
@@ -680,7 +674,6 @@ func Example_zone() {
 	// constraints: []
 	// zone ls
 	// .default
-	// system
 	// system.jobs
 	// t
 	// zone set t.f --file=./testdata/zone_range_max_bytes.yaml
@@ -692,7 +685,6 @@ func Example_zone() {
 	// constraints: []
 	// zone ls
 	// .default
-	// system
 	// system.jobs
 	// t
 	// t.f
@@ -700,7 +692,6 @@ func Example_zone() {
 	// DROP DATABASE
 	// zone ls
 	// .default
-	// system
 	// system.jobs
 }
