+++ pkg/storage/allocator_scorer.go
@@ -430,7 +430,7 @@ func allocateCandidates(
 ) candidateList {
 	var candidates candidateList
 	for _, s := range sl.stores {
-		if storeHasReplica(s.StoreID, existing) {
+		if StoreHasReplica(s.StoreID, existing) {
 			continue
 		}
 		constraintsOK, necessary := allocateConstraintsCheck(s, constraints)
@@ -872,7 +872,7 @@ func shouldRebalance(
 					sl.candidateWritesPerSecond.mean)
 				continue
 			}
-			if storeHasReplica(desc.StoreID, rangeInfo.Desc.Replicas) {
+			if StoreHasReplica(desc.StoreID, rangeInfo.Desc.Replicas) {
 				continue
 			}
 			log.VEventf(ctx, 2,
@@ -927,9 +927,9 @@ func shouldRebalanceNoStats(
 	return false
 }
 
-// storeHasReplica returns true if the provided NodeID contains an entry in
+// StoreHasReplica returns true if the provided NodeID contains an entry in
 // the provided list of existing replicas.
-func storeHasReplica(storeID roachpb.StoreID, existing []roachpb.ReplicaDescriptor) bool {
+func StoreHasReplica(storeID roachpb.StoreID, existing []roachpb.ReplicaDescriptor) bool {
 	for _, r := range existing {
 		if r.StoreID == storeID {
 			return true
@@ -995,7 +995,7 @@ func analyzeConstraints(
 			// is a much more stable failure state than frantically moving everything
 			// off such a node.
 			store, ok := getStoreDescFn(repl.StoreID)
-			if !ok || subConstraintsCheck(store, subConstraints.Constraints) {
+			if !ok || SubConstraintsCheck(store, subConstraints.Constraints) {
 				result.satisfiedBy[i] = append(result.satisfiedBy[i], store.StoreID)
 				result.satisfies[store.StoreID] = append(result.satisfies[store.StoreID], i)
 			}
@@ -1025,7 +1025,7 @@ func allocateConstraintsCheck(
 	}
 
 	for i, constraints := range analyzed.constraints {
-		if constraintsOK := subConstraintsCheck(store, constraints.Constraints); constraintsOK {
+		if constraintsOK := SubConstraintsCheck(store, constraints.Constraints); constraintsOK {
 			valid = true
 			matchingStores := analyzed.satisfiedBy[i]
 			if len(matchingStores) < int(constraints.NumReplicas) {
@@ -1098,7 +1098,7 @@ func rebalanceFromConstraintsCheck(
 	// of constraints with NumReplicas set to 0. This is meant to be enforced in
 	// the config package.
 	for i, constraints := range analyzed.constraints {
-		if constraintsOK := subConstraintsCheck(store, constraints.Constraints); constraintsOK {
+		if constraintsOK := SubConstraintsCheck(store, constraints.Constraints); constraintsOK {
 			valid = true
 			matchingStores := analyzed.satisfiedBy[i]
 			if len(matchingStores) < int(constraints.NumReplicas) ||
@@ -1134,17 +1134,17 @@ func constraintsCheck(store roachpb.StoreDescriptor, constraints []config.Constr
 	}
 
 	for _, subConstraints := range constraints {
-		if constraintsOK := subConstraintsCheck(store, subConstraints.Constraints); constraintsOK {
+		if constraintsOK := SubConstraintsCheck(store, subConstraints.Constraints); constraintsOK {
 			return true
 		}
 	}
 	return false
 }
 
-// subConstraintsCheck checks a store against a single set of constraints (out
+// SubConstraintsCheck checks a store against a single set of constraints (out
 // of the possibly numerous sets that apply to a range), returning true iff the
 // store matches the constraints.
-func subConstraintsCheck(store roachpb.StoreDescriptor, constraints []config.Constraint) bool {
+func SubConstraintsCheck(store roachpb.StoreDescriptor, constraints []config.Constraint) bool {
 	for _, constraint := range constraints {
 		if !config.StoreMatchesConstraint(store, constraint) {
 			return false
